+++ src/com/algoTrader/subscriber/SetExitValueSubscriber.java
@@ -1,12 +1,10 @@
 package com.algoTrader.subscriber;
 
-import java.math.BigDecimal;
-
 import com.algoTrader.ServiceLocator;
 
 public class SetExitValueSubscriber {
 
-    public void update(int positionId, BigDecimal exitValue) {
+    public void update(int positionId, double exitValue) {
 
         ServiceLocator.instance().getActionService().setExitValue(positionId, exitValue);
     }
